break up 'main purpose components' to Header.js and Footer.js
